# Speed up opening cached light curve files by a factor ~5x