As reported in #914, obtaining light curves via `download_all()` was surprisingly slow, even what all the files are available in the local cache.

Using `line_profiler`, I determined the primary culprit to be the fact that `astroquery.mast.Observations.download_products()` always submits a HTTP request to MAST to determine the length of the file prior to checking if the file already exists in the local cache.  This PR modifies `download()` and `download_all()` to skip that request.

This PR also includes a few easy speed-ups inside`lightkurve.io.generic.read_generic_lightcurve()`.

As a result, the time required to open 10 light curves (cf. example in #914) dropped from ~8.5s to ~1.5s, which is a ~5x improvement:

**Before:**
```python
>>> %time collection = search.download_all()
CPU times: user 4.55 s, sys: 169 ms, total: 4.72 s
Wall time: 8.71 s
```

**After:**
```python
>>> %time collection = search.download_all()
CPU times: user 1.26 s, sys: 100 ms, total: 1.36 s
Wall time: 1.56 s
```