@barentsen 2020-12-02
I expect to merge this PR after fixing the Windows CI failure and adding a unit test.